Insurance-Claim-Help-API-docker-vps-deploy/requirements.txt
asgiref==3.11.1
charset-normalizer==3.4.6
Django==6.0.3
django-cors-headers==4.9.0
django-unfold==0.63.0
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
gunicorn==23.0.0
mysql-connector-python==9.6.0
pillow==12.1.1
PyJWT==2.12.1
python-dotenv==1.2.2
reportlab==4.4.10
sqlparse==0.5.5
whitenoise==6.9.0
